+++ app/main/db/queries.py
@@ -1,8 +1,10 @@
 import uuid
 
-from flask import current_app
+from flask import current_app, session
 from sqlalchemy import DATE, Text, and_, func, or_
 
+from app.main.authorize.ayr_user import AYRUser
+from app.main.authorize.keycloak_manager import get_user_groups
 from app.main.authorize.permissions_helpers import (
     validate_body_user_groups_or_404,
 )
@@ -24,7 +26,6 @@ def browse_data(
     consignment_id=None,
     date_range=None,
     date_filter_field=None,
-    user_transferring_body=None,
 ):
     if transferring_body_id:
         body = Body.query.get_or_404(transferring_body_id)
@@ -46,7 +47,10 @@ def browse_data(
             date_filter_field=date_filter_field,
         )
     else:
-        browse_query = _build_browse_everything_query(user_transferring_body)
+        ayr_user = AYRUser(get_user_groups(session.get("access_token")))
+        browse_query = _build_browse_everything_query(
+            ayr_user.transferring_body
+        )
 
     if not consignment_id and date_range:
         dt_range = validate_date_range(date_range)
